+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-
-import 'package:not_too_shabby/app.dart';
+import 'package:not_too_shabby/view/app.dart';
 
 void main() => runApp(NotTooShabbyApp());