+++ lib/model/youtube_api_key.dart
@@ -1,0 +1,10 @@
+class YoutubeApiKey {
+  final String value;
+
+  YoutubeApiKey(
+    this.value,
+  );
+
+  YoutubeApiKey.fromJson(Map<String, dynamic> json)
+      : value = json['youtubeApiKey'];
+}